move testTheme into test package
Thus it is possible to use `test.WithTestTheme` outside the widget package.
